# issue 158: Add SmallState