Is an optional reduced size state.
Fixes #158