@arthurprs 2024-03-23
I performed a quick benchmark with runtime-rng enabled (the most common feature I suppose) and the time to hash a 10 byte string goes from 2.5ns to 12.3ns, which puts it in siphash territory.

Do you think we could take some liberties with the creation of RandomState from SmallState, like

```
        let [a, _] = get_fixed_seeds();
        AHasher::from_random_state(&RandomState {
            k0: self.key as u64 ^ a[0],
            k1: self.key as u64 ^ a[1],
            k2: self.key as u64 ^ a[2],
            k3: self.key as u64 ^ a[3],
            _h: PhantomData::<T>,
        })
```

This brings it down from 12ns to 4ns