@tkaitchuck 2024-03-25
@arthurprs The code you have there won't work. But I have tried to speed it up. Give it another try.