@arthurprs 2024-03-25
I got 7ns for 3e14ceb and 6.3ns for 1e94eac.
For context, it's 2.5ns for the regular ahash and 10.3ns for sip13.

I think it's good now, unless you can squeeze more out of it :rocket: