@tkaitchuck 2024-03-26
@arthurprs Is [e2ae7b7](https://github.com/tkaitchuck/aHash/pull/226/commits/e2ae7b75674a29cc77fd764afab0a5efc4522f5c) any faster on your system? It *should* be, but it my tests it appears to be, but only by about the same margin as random noise.