@arthurprs 2024-03-26
I can confirm the latest commit is 12% faster than the previous (~5ns).